requirements.txt
onnxruntime
flask
waitress
pillow
numpy
